--- a/src/.babelrc
+++ b/src/.babelrc
@@ -1,8 +1,0 @@
-{
-  "plugins": [
-    "syntax-jsx",
-    "transform-es2015-destructuring",
-    "transform-object-rest-spread",
-    "transform-react-jsx"
-  ]
-}