@feross 2016-09-20
Fixes https://github.com/feross/webtorrent-desktop/issues/940