@Flet 2016-09-20
Its probably negligible, but is there a big difference in time running integration tests?