@feross 2016-09-21
> is there a big difference in time running integration tests?

I think the `npm run build` happens only once in the integration tests. But if ES5 is faster than ES6, then maybe we'll see that in the time it takes integration tests to complete...