@Flet 2016-09-21
>ES5 is faster than ES6, then maybe we'll see that in the time it takes integration tests to complete

Yeah, that's what I'm curious about :)